+++ src/modules/logo/option.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "common/option.h"
+
+typedef struct FFLogoOptions
+{
+    FFModuleArgs moduleArgs;
+} FFLogoOptions;
+
+static_assert(sizeof(FFLogoOptions) <= FF_OPTION_MAX_SIZE, "FFLogoOptions size exceeds maximum allowed size");